feat(politics-tracker): modify section toggle behavior
